Decouple services from strategies
