+++ common/base_strategy_config.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class BaseStrategyConfig:
+    """모든 전략 공통 설정."""
+
+    use_fixed_qty: bool = True
