+++ services/newhigh_strategy_coverage_backtest_service.py
@@ -9,7 +9,6 @@ from typing import Any, Callable
 from common.trade_journal_schema import normalize_backtest_decision
 from services.backtest_replay_adapter import StockQueryBacktestReplayService
 from services.backtest_replay_context import BacktestMarketClock
-from strategies.debug.strategy_debug_runner import StrategyDebugRunner
 
 
 _DEFAULT_STRATEGY_KEYS = (
@@ -66,6 +65,7 @@ class NewHighStrategyCoverageBacktestService:
         backtest_journal_repository: Any,
         program_provider: Any | None = None,
         strategy_factory: Callable[..., Any] | None = None,
+        debug_runner_factory: Callable[..., Any] | None = None,
         env: Any | None = None,
         logger: logging.Logger | None = None,
     ) -> None:
@@ -77,6 +77,9 @@ class NewHighStrategyCoverageBacktestService:
         self._repo = backtest_journal_repository
         self._program_provider = program_provider
         self._strategy_factory = strategy_factory or self._default_strategy_factory
+        if debug_runner_factory is None:
+            raise ValueError("debug_runner_factory must be provided by the composition root")
+        self._debug_runner_factory = debug_runner_factory
         self._env = env
         self._logger = logger or logging.getLogger(__name__)
         self._debug_logger = logger if isinstance(logger, logging.Logger) else logging.getLogger(__name__)
@@ -161,7 +164,7 @@ class NewHighStrategyCoverageBacktestService:
                 logger=self._debug_logger,
             )
             strategy_name = str(getattr(strategy, "name", strategy_key) or strategy_key)
-            report = await StrategyDebugRunner(
+            report = await self._debug_runner_factory(
                 strategy,
                 self._debug_logger,
                 target_date=target_date,
